requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dateutil
python-dotenv
requests
